+++ src/Core/Tests/ContentManager/Collection/CollectionManagerTest.php
@@ -20,14 +20,14 @@ class CollectionManagerTest extends \PHPUnit_Framework_TestCase
     public function testGetCollectionItemCollection()
     {
         $cm = new CollectionManager();
-        $this->assertInstanceOf('Yosymfony\Spress\Core\ContentManager\Collection\ColectionItemCollection', $cm->getCollectionItemCollection());
+        $this->assertInstanceOf('Yosymfony\Spress\Core\Support\Collection', $cm->getCollectionItemCollection());
     }
 
     public function testGetCollectionForItems()
     {
         $cm = new CollectionManager();
-        $cm->getCollectionItemCollection()->add(new Collection('events', 'events'));
-        $cm->getCollectionItemCollection()->add(new Collection('books', 'books'));
+        $cm->getCollectionItemCollection()->add('events',new Collection('events', 'events'));
+        $cm->getCollectionItemCollection()->add('books', new Collection('books', 'books'));
 
         $item = new Item('Test of content', 'events/event-1.html');
         $item->setPath('events/event-1.html', Item::SNAPSHOT_PATH_RELATIVE);